Ported changes from ZDoom revision 2233-2249 and from GZDoom revision 755-757: - added a CrushPainSound actor property for Strife. - fixed memory leaks in SBARINFO and WAD loading code. - added GetBloodColor and GetBloodType inline functions to AActor to wrap the GetMeta calls used for this. - Changed A_SetUserVar and A_SetUserArray so they affect the actor that called it, which is not necessarily "self". The only visible change from this should be that inventory items now set their own variables and not their owners'. - Fixed: TMap::DelKey failed if the key's main position was nil, because it tried checking for it at the "next" position, which is an invalid pointer in that case. - Keep all damage factors in the table, even those that are 1.0. - 256 is a valid pain chance, so clamp to that, not 255. - P_ZMovement() temporarily disables jumping after a landing. Don't do this if the jump timer is already running or for short falls (e.g. along the edges of slopes, since the slope floorz calculation is pretty crappy.) - fixed: Level redirection checked the wrong level. - fix division by 0. - fixed: P_SpawnMapThing always reduced the angular precision to 45 degrees. - removed AngleIncrements because it's not really useful. - fixed: IDBEHOLD altered the item counter. - Use normal texture animation for the main menu cursors. This required updating animations all the time and not just when inside a level. - fixed: Explosions directly under a water surface would not hurt any actor directly above this surface. - cleaned up P_CheckSight flag handling.

+++ src/p_interaction.cpp
@@ -1664,7 +1664,7 @@ void P_DamageMobj (AActor *target, AActor *inflictor, AActor *source, int damage
 		painchance = target->PainChance;
 		if (pc != NULL)
 		{
-			BYTE * ppc = pc->CheckKey(mod);
+			int *ppc = pc->CheckKey(mod);
 			if (ppc != NULL)
 			{
 				painchance = *ppc;
